support of quantity as 'double' datatype

+++ lib/src/pages/create_purchase.dart
@@ -202,7 +202,7 @@ class _CreatePurchaseState extends State<CreatePurchase> {
     final orderItems = temp
         .map((e) => OrderItemInput(
               product: e,
-              quantity: tempMap["${e.id}"],
+              quantity: tempMap["${e.id}"].toDouble(),
               price: 0,
             ))
         .toList();
